python/requirements.txt
google-genai
numpy
numba
